Merge dependencies jobs
GitHub disables scheduled jobs if there hasn't been activity on the
repository for the last 60 days. This is likely to happen a lot here
since we will probably only update it after breaking ruma releases that
should happen every 6 months or so, so we remove the schedule job.

Since both jobs are now always running with the same conditions, we can
just merge them.